requirements.txt
tweepy==4.15.0
dotenv==0.9.9
google-genai==1.7.0
playwright==1.51.0